+++ changelog.md
@@ -1,3 +1,11 @@
+# Version 2024.11.3 (2024-11-18)
+
+- Add interface(id) to performance log message
+- Add interfaces_requiring_periodic_refresh to config
+- Add periodic data refresh to CentralUnitChecker for some interfaces
+- Maintain data_cache by interface
+- Reduce MAX_CACHE_AGE to 15s
+
 # Version 2024.11.2 (2024-11-17)
 
 - Add get_data_point_path to central